Fix loading filesize coordinator from wrong place
